Fixes for compilation on Linux.
git-svn-id: svn://localhost@1252 8062f311-0dae-4547-b526-b8ab9ac864a5

+++ GDCpp/GDCpp/BuiltinExtensions/CommonInstructionsExtension.cpp
@@ -212,7 +212,7 @@ CommonInstructionsExtension::CommonInstructionsExtension()
         {
             virtual std::string GenerateCode(gd::Instruction & instruction, gd::EventsCodeGenerator & codeGenerator, gd::EventsCodeGenerationContext & parentContext)
             {
-                unsigned int uniqueId = (unsigned int)&instruction;
+                size_t uniqueId = (size_t)&instruction;
                 return "conditionTrue = runtimeContext->TriggerOnce("+ToString(uniqueId)+");\n";
             };
         };
